Emit errors instead of throwing on parse errors
If an error is thrown while creating a Request object, such as when an
unknown resource record type is requested, there's no way for the client
to handle it. The error is thrown before the request is emitted and the
handler is called. This change leaves the underlying error throwing as
it is but catches the error in the server and emits them instead.