Emit socket errors from the server
Any errors that might be emitted from the socket rather than the server
can't be handled by the client. This listens for errors on the connection
(and request and response) and emits them from the server.

In order to do that, the Request and Response objects (which inherit from
DNSMessage) must actually be EventEmitters. They often call emit, but only
in rare code paths. Those code paths would throw because emit wasn't a function.